refactor: move get_command_from_path where it belongs
