+++ pyrevitlib/pyrevit/script.py
@@ -32,7 +32,8 @@ from pyrevit import versionmgr
 from pyrevit.labs import PyRevit
 from pyrevit.revit import ui
 import pyrevit.extensions as exts
-
+from pyrevit.extensions.parser import parse_comp_dir
+from pyrevit.extensions.genericcomps import GenericUICommand
 
 # suppress any warning generated by native or third-party modules
 warnings.filterwarnings("ignore")
@@ -54,8 +55,23 @@ def get_info():
         (pyrevit.extensions.genericcomps.GenericUICommand):
             Command info object
     """
-    from pyrevit.extensions.extensionmgr import get_command_from_path
     return get_command_from_path(EXEC_PARAMS.command_path)
+
+
+def get_command_from_path(comp_path):
+    """Returns a pyRevit command object from the given bundle directory.
+
+    Args:
+        comp_path (str): Full directory address of the command bundle
+
+    Returns:
+        (genericcomps.GenericUICommand): A subclass of pyRevit command object.
+    """
+    cmds = parse_comp_dir(comp_path, GenericUICommand)
+    if cmds:
+        return cmds[0]
+
+    return None
 
 
 def get_script_path():
